Add files via upload
In my master’s study, I independently developed a creative function learning algorithm to summarize the experimental data. I intuitively created a simple formula which could accurately present more than 1GB of velocity data and make predictions for the missing experimental conditions.  The formula was the result of more than 200 non-linear regressions that was later used in the numerical simulations to predict the collapse of the volcanic column. The regression formula encompassed the entrained velocity data as a function of the overpressure ratio, the axial location from the jet, and the density ratio. Researchers have cited the published work more than 18 times. "Volcano.mat" is a dataset and py file is the function learning code.